requirements.txt
requests
python-dotenv
solana==0.35.0
solders==0.21.0
supabase